Manage sticky navbar using react helmet
So the reason for this is that because of rehydration, the
useLayoutEffect will only run *after* the page is being displayed,
because most of everything else has been rendered server side.

We solve this using React helmet to insert some html data properties
into the body element, and then using those to create the required
styles. Because react helmet works at compile and runtime, it works on
first load and all subsequent navigation.

It may seem weird to use html data properties, especially when we could
just use style attributes, however, we choose to avoid those as other
React Helmet components in the future may want to use a style attribute
as well, overriding the existing one. Keeping them as a data properties
ensures their uniqueness, even at a price of some slightly more complex
css.