Use a sticky nav (instead of the builtin bulma fixed one)
Yes, this was fiddly to eventually decide on. It's not ideal, as sticky
positioning is perfectly supported on everything, however, for now it'll
work. Also, even it fails, the default fallback will just be to not have
a sticky header at all, so it'll be fine.

The official bulma solution is really difficult to pull off, mostly
because, it's an incredible pain in the ass to add classnames to the
body or html elements well. It can be done, but with a bunch of annoying
caveats.