Give non standard posts and pages more support
Page titles and dates are conditionally rendered now, which means that
they can be ommitted in the front matter if neccessary. Additionally, we
support the new frontmatter tag "hidden" which allows a post to be
hidden from the global blog post listing.

Finally, we use GraphQL typing to ensure that the frontmatter fields are
properly documented and listed.